third_party/WebKit/LayoutTests/media/video-playing-and-pause.html

<!DOCTYPE HTML>
<html>
<body>
    <video id="testVideo" controls></video>
    <div id="log"></div>
    <p>Test that pausing the media element in "playing" event handler pauses the media element.</p>
    <script src="./w3c-media-utils.js"></script>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script>
        var pauseTest = new async_test("Verifies pause actually pauses.");
        var v = document.getElementById("testVideo");
        v.src = getVideoURI("test");
        waitForEventAndRunStep("timeupdate", v);
        waitForEventAndRunStep("playing", v, function ()
        {
            v.pause();
        });
        // function to be run when "pause" fires.
        var pauseCheck = function()
        {
            assert_true(v.paused, "Video is paused.");
            pauseTest.done();
        }
        waitForEventAndRunStep("pause", v, pauseCheck, pauseTest);
        v.play();
     </script>
  </body>
</html>
